+++ test/socketio/context_socket.io.test.js
@@ -16,6 +16,7 @@ const {
   waitForNoEvent,
   enterContext,
   exitContext,
+  resetContexts,
 } = require("../_env/util/socket.io");
 
 describe("Context", () => {
@@ -51,7 +52,7 @@ describe("Context", () => {
     expect(eventResult.text).toBe("test1");
     await eventNoResultPromise;
 
-    await enterContext(socketOther, ID);
+    await enterContext(socketOther, [ID]);
     await wait();
     eventResultPromise = waitForEvent(socket, "customContextIncludeEvent");
     const eventResultOtherPromise = waitForEvent(socketOther, "customContextIncludeEvent");
@@ -72,8 +73,7 @@ describe("Context", () => {
     expect(eventResult.text).toBe("test1");
     await eventNoResultPromise;
 
-    await exitContext(socket, ID);
-    await exitContext(socketOther, ID);
+    await resetContexts(socket);
     await wait();
     eventNoResultPromise = waitForNoEvent(socket, "customContextIncludeEvent");
     const eventNoOtherResultPromise = waitForNoEvent(socketOther, "customContextIncludeEvent");
@@ -109,7 +109,7 @@ describe("Context", () => {
     const eventResultOther = await eventResultOtherPromise;
     expect(eventResultOther.text).toBe("test1");
 
-    await exitContext(socketOther, context);
+    await exitContext(socketOther, [context]);
     await wait();
     eventResultPromise = waitForEvent(socket, "customContextIncludeStaticEvent");
     eventNoResultPromise = waitForNoEvent(socketOther, "customContextIncludeStaticEvent");